+++ java/src15/org/garret/perst/impl/RtreePage.java
@@ -51,11 +51,11 @@ public class RtreePage extends Persistent {
         if (--level != 0) { 
             // not leaf page
             int i, mini = 0;
-            int minIncr = Integer.MAX_VALUE;
-            int minArea = Integer.MAX_VALUE;
+            long minIncr = Long.MAX_VALUE;
+            long minArea = Long.MAX_VALUE;
             for (i = 0; i < n; i++) { 
-                int area = b[i].area();
-                int incr = Rectangle.joinArea(b[i], r) - area;
+                long area = b[i].area();
+                long incr = Rectangle.joinArea(b[i], r) - area;
                 if (incr < minIncr) { 
                     minIncr = incr;
                     minArea = area;
@@ -159,9 +159,9 @@ public class RtreePage extends Persistent {
 
     final RtreePage splitPage(Branch br) { 
         int i, j, seed0 = 0, seed1 = 0;
-        int[]  rectArea = new int[card+1];
-        int    waste;
-        int    worstWaste = Integer.MIN_VALUE;
+        long[] rectArea = new long[card+1];
+        long   waste;
+        long   worstWaste = Long.MIN_VALUE;
         //
         // As the seeds for the two groups, find two rectangles which waste 
         // the most area if covered by a single rectangle.
@@ -184,7 +184,7 @@ public class RtreePage extends Persistent {
         }       
         byte[] taken = new byte[card];
         Rectangle group0, group1;
-        int       groupArea0, groupArea1;
+        long      groupArea0, groupArea1;
         int       groupCard0, groupCard1;
         RtreePage pg;
 
@@ -213,11 +213,11 @@ public class RtreePage extends Persistent {
                && groupCard1 < card + 1 - minFill)
         {
             int betterGroup = -1, chosen = -1;
-            int biggestDiff = -1;
+            long biggestDiff = -1;
             for (i = 0; i < card; i++) { 
                 if (taken[i] == 0) { 
-                    int diff = (Rectangle.joinArea(group0, b[i]) - groupArea0)
-                             - (Rectangle.joinArea(group1, b[i]) - groupArea1);
+                    long diff = (Rectangle.joinArea(group0, b[i]) - groupArea0)
+                              - (Rectangle.joinArea(group1, b[i]) - groupArea1);
                     if (diff > biggestDiff || -diff > biggestDiff) { 
                         chosen = i;
                         if (diff < 0) {
